[IMP] html_editor: move video plugins
task-5318096
