+++ addons/html_editor/static/src/main/media/video/video_plugin.js
