+++ addons/html_editor/static/src/main/media/video/youtube_plugin.js
