+++ addons/html_editor/static/src/plugin_sets.js
@@ -56,8 +56,8 @@ import { TableUIPlugin } from "./main/table/table_ui_plugin";
 import { TabulationPlugin } from "./main/tabulation_plugin";
 import { TextDirectionPlugin } from "./main/text_direction_plugin";
 import { ToolbarPlugin } from "./main/toolbar/toolbar_plugin";
-import { VideoPlugin } from "./main/media/video_plugin";
-import { YoutubePlugin } from "./main/youtube_plugin";
+import { VideoPlugin } from "./main/media/video/video_plugin";
+import { YoutubePlugin } from "./main/media/video/youtube_plugin";
 import { PlaceholderPlugin } from "./main/placeholder_plugin";
 import { CollaborationOdooPlugin } from "./others/collaboration/collaboration_odoo_plugin";
 import { CollaborationPlugin } from "./others/collaboration/collaboration_plugin";
